requirements.txt
boto3
requests
discord
python-dotenv